Rename parquet_helpers.hpp to parquet_stats_file_helpers.hpp
