+++ src/io/parquet_stats_file_writer_impl.hpp
@@ -39,7 +39,7 @@ NCHG_DISABLE_WARNING_POP
 #include <stdexcept>
 #include <string>
 
-#include "nchg/parquet_helpers.hpp"
+#include "nchg/parquet_stats_file_helpers.hpp"
 
 namespace nchg {
